+++ packages/tap/src/core/ResourceHandle.ts
@@ -1,4 +1,4 @@
-import { ResourceElement, Unsubscribe } from "./types";
+import { ResourceElement } from "./types";
 import {
   createResourceFiber,
   unmountResource,
@@ -11,62 +11,69 @@ import { tapState } from "../hooks/tap-state";
 import { tapMemo } from "../hooks/tap-memo";
 import { tapInlineResource } from "../hooks/tap-inline-resource";
 import { tapEffect } from "../hooks/tap-effect";
+import { resource } from "./resource";
 
-export interface ResourceHandle<R, P> {
-  getState(): R;
-  subscribe(callback: () => void): Unsubscribe;
-  updateInput(props: P): void;
-  flushSync(): void;
-  dispose(): void;
+export namespace createResource {
+  export type Unsubscribe = () => void;
+
+  export interface Handle<R, P> {
+    getState(): R;
+    subscribe(callback: () => void): Unsubscribe;
+    updateInput(props: P): void;
+    flushSync(): void;
+    dispose(): void;
+  }
 }
 
-const HandleWrapperResource = <R, P>({
-  element,
-  onUpdateInput,
-  onFlushSync,
-  onDispose,
-}: {
-  element: ResourceElement<R, P>;
-  onUpdateInput: () => void;
-  onFlushSync: () => void;
-  onDispose: () => void;
-}): ResourceHandle<R, P> => {
-  const [props, setProps] = tapState(element.props);
-  const value = tapInlineResource({ type: element.type, props });
-  const subscribers = tapRef(new Set<() => void>()).current;
-  const valueRef = tapRef(value);
+const HandleWrapperResource = resource(
+  <R, P>({
+    element,
+    onUpdateInput,
+    onFlushSync,
+    onDispose,
+  }: {
+    element: ResourceElement<R, P>;
+    onUpdateInput: () => void;
+    onFlushSync: () => void;
+    onDispose: () => void;
+  }): createResource.Handle<R, P> => {
+    const [props, setProps] = tapState(element.props);
+    const value = tapInlineResource({ type: element.type, props });
+    const subscribers = tapRef(new Set<() => void>()).current;
+    const valueRef = tapRef(value);
 
-  tapEffect(() => {
-    if (value !== valueRef.current) {
-      valueRef.current = value;
-      subscribers.forEach((callback) => callback());
-    }
-  });
+    tapEffect(() => {
+      if (value !== valueRef.current) {
+        valueRef.current = value;
+        subscribers.forEach((callback) => callback());
+      }
+    });
 
-  const handle = tapMemo(
-    () => ({
-      getState: () => valueRef.current,
-      subscribe: (callback: () => void) => {
-        subscribers.add(callback);
-        return () => subscribers.delete(callback);
-      },
-      updateInput: (props: P) => {
-        onUpdateInput();
-        setProps(() => props);
-      },
-      flushSync: onFlushSync,
-      dispose: onDispose,
-    }),
-    [],
-  );
+    const handle = tapMemo(
+      () => ({
+        getState: () => valueRef.current,
+        subscribe: (callback: () => void) => {
+          subscribers.add(callback);
+          return () => subscribers.delete(callback);
+        },
+        updateInput: (props: P) => {
+          onUpdateInput();
+          setProps(() => props);
+        },
+        flushSync: onFlushSync,
+        dispose: onDispose,
+      }),
+      [],
+    );
 
-  return handle;
-};
+    return handle;
+  },
+);
 
 export const createResource = <R, P>(
   element: ResourceElement<R, P>,
   delayMount = false,
-): ResourceHandle<R, P> => {
+): createResource.Handle<R, P> => {
   let isMounted = !delayMount;
   const props = {
     element,