+++ packages/tap/src/core/resource.ts
@@ -1,15 +1,16 @@
-import { ResourceFn, ResourceElement, Resource } from "./types";
-import { fnSymbol } from "./getResourceFn";
-
+import { ResourceElement, Resource } from "./types";
+import { fnSymbol } from "./callResourceFn";
+export function resource<R>(fn: () => R): () => ResourceElement<R, undefined>;
+export function resource<R, P>(fn: (props: P) => R): Resource<R, P>;
 export function resource<R, P = undefined>(
-  fn: ResourceFn<R, P>,
+  fn: (props: P) => R,
 ): Resource<R, P> {
   const type = (props?: P, options?: { key?: string | number }) => {
     return {
       type,
-      props,
+      props: props!,
       ...(options?.key !== undefined && { key: options.key }),
-    } as ResourceElement<R, P>;
+    } satisfies ResourceElement<R, P>;
   };
 
   type[fnSymbol] = fn;