+++ packages/tap/src/hooks/tap-inline-resource.ts
@@ -1,6 +1,6 @@
 import { ResourceElement } from "../core/types";
-import { getResourceFn } from "../core/getResourceFn";
+import { callResourceFn } from "../core/callResourceFn";
 
 export function tapInlineResource<R, P>(element: ResourceElement<R, P>): R {
-  return getResourceFn(element.type)(element.props);
+  return callResourceFn(element.type, element.props);
 }