+++ packages/tap/src/hooks/tap-ref.ts
@@ -1,12 +1,14 @@
 import { tapState } from "./tap-state";
 
-export interface RefObject<T> {
-  current: T;
+export namespace tapRef {
+  export interface RefObject<T> {
+    current: T;
+  }
 }
 
-export function tapRef<T>(initialValue: T): RefObject<T>;
-export function tapRef<T = undefined>(): RefObject<T | undefined>;
-export function tapRef<T>(initialValue?: T): RefObject<T | undefined> {
+export function tapRef<T>(initialValue: T): tapRef.RefObject<T>;
+export function tapRef<T = undefined>(): tapRef.RefObject<T | undefined>;
+export function tapRef<T>(initialValue?: T): tapRef.RefObject<T | undefined> {
   const [state] = tapState(() => ({
     current: initialValue,
   }));