+++ packages/tap/src/react/use-resource.ts
@@ -1,6 +1,5 @@
 import { useEffect, useLayoutEffect, useMemo, useState } from "react";
 import { ResourceElement } from "../core/types";
-import { getResourceFn } from "../core/getResourceFn";
 import {
   createResourceFiber,
   unmountResource,
@@ -18,7 +17,7 @@ const useIsomorphicLayoutEffect = shouldAvoidLayoutEffect
 export function useResource<R, P>(element: ResourceElement<R, P>): R {
   const [, rerender] = useState({});
   const fiber = useMemo(
-    () => createResourceFiber(getResourceFn(element.type), () => rerender({})),
+    () => createResourceFiber(element.type, () => rerender({})),
     [element.type, rerender],
   );